Region 클래스를 Cuboid 클래스로 변경
