--- src/main/java/net/teamuni/shootingtest/Region.java
@@ -1,90 +1,0 @@
-package net.teamuni.shootingtest;
-
-import org.bukkit.Location;
-import org.bukkit.World;
-
-public class Region {
-    private final World world;
-    private final int minX;
-    private final int maxX;
-    private final int minZ;
-    private final int maxZ;
-
-    public Region(Location loc1, Location loc2) {
-        this(loc1.getWorld(), loc1.getBlockX(), loc1.getBlockZ(), loc2.getBlockX(), loc2.getBlockZ());
-    }
-
-    public Region(World world, int x1, int z1, int x2, int z2) {
-        this.world = world;
-
-        minX = Math.min(x1, x2);
-        minZ = Math.min(z1, z2);
-        maxX = Math.max(x1, x2);
-        maxZ = Math.max(z1, z2);
-    }
-
-    public World getWorld() {
-        return world;
-    }
-
-    public int getMinX() {
-        return minX;
-    }
-
-    public int getMinZ() {
-        return minZ;
-    }
-
-    public int getMaxX() {
-        return maxX;
-    }
-
-    public int getMaxZ() {
-        return maxZ;
-    }
-
-    public boolean contains(Region region) {
-        return region.getWorld().equals(world) &&
-                region.getMinX() >= minX && region.getMaxX() <= maxX &&
-                region.getMinZ() >= minZ && region.getMaxZ() <= maxZ;
-    }
-
-    public boolean contains(Location location) {
-        return contains(location.getBlockX(), location.getBlockZ());
-    }
-
-    public boolean contains(int x, int z) {
-        return x >= minX && x <= maxX &&
-                z >= minZ && z <= maxZ;
-    }
-
-    public boolean overlaps(Region region) {
-        return region.getWorld().equals(world) &&
-                !(region.getMinX() > maxX || region.getMinZ() > maxZ ||
-                        minZ > region.getMaxX() || minZ > region.getMaxZ());
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj) {
-            return true;
-        }
-        if (!(obj instanceof final Region other)) {
-            return false;
-        }
-        return world.equals(other.world)
-                && minX == other.minX
-                && minZ == other.minZ
-                && maxX == other.maxX
-                && maxZ == other.maxZ;
-    }
-
-    @Override
-    public String toString() {
-        return "Region[world:" + world.getName() +
-                ", minX:" + minX +
-                ", minZ:" + minZ +
-                ", maxX:" + maxX +
-                ", maxZ:" + maxZ + "]";
-    }
-}
